python-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
